Saves

+++ proyecto[Practica01]/Dominio/Articulo.cs
@@ -11,7 +11,7 @@ namespace proyecto_Practica01_.Dominio
         public int id_articulo {  get; set; }
         public string Nombre {  get; set; }
 
-        public int PrecioUnitario { get; set; }
+        public decimal PrecioUnitario { get; set; }
 
         public override string ToString()
         {
